[probably bugfix?] Compiler complains about MMX
Noticed that couldn't compile SSR because of some bug related to
xmmintrin.h, mmmintrin.h and tmmintrin.h files. After looking at the
code, noticed that an #error was being triggered because some __MMX__
constant wasn't defined.

Found that both AV/FastScaler_Convert.cpp and AV/FastScaler_Scale.cpp
defined constants related to CPU features such as SSE or SSE2, but
not MMX, which is another CPU feature, so I defined __MMX__ too.

I don't know if leaving out __MMX__ was intentional or not or even
if adding __MMX__ is completely unrelated to the code but it seems
to work as now it compiles again. I've tested it by doing some test
recordings and I don't seem to receive any errors so I suppose it's a
bugfix.