Added fragmented recording backend.
It is currently disabled and not exposed by the GUI. It turns out that existing tools cannot recombine the fragments in a glitch-free way, so this is not very useful yet. I plan to add code to handle the recombining process myself. Sadly this can only be done with low-level libavcodec/libavformat code, not with simple scripts.