Use the X11 Xinerama extension to get screen resolutions in X11Input.
This replaces the Qt-based system, keeping Qt code out of the backend and also avoiding the whole high DPI scaling issue.

The downside is that we can't update the resolution in real-time anymore, since that would require XRandR (which is not always available). But that was never really supported anyway.