Add function to open a folder using a file browser
I've tested this with local files as well as with an SFTP drive. Both open the directory with Nautilus, even though the Qt documentation says that anything that is not a local file should be opened with a web browser. But this is more desirable than what Qt documents in any case.